Moved Buffer, Accessor, Vertex to own files
