+++ blendergltf.py
@@ -1,16 +1,15 @@
 import base64
-import collections
 import functools
-import math
 import os
 import shutil
 import struct
 import zlib
-
 import bpy
 import gpu
 import mathutils
 
+from .types.Buffer import Buffer
+from .types.Vertex import Vertex
 
 if '_IMPORTED' not in locals():
     _IMPORTED = True
@@ -60,302 +59,6 @@ PROFILE_MAP = {
     'WEB': {'api': 'WebGL', 'version': '1.0.3'},
     'DESKTOP': {'api': 'OpenGL', 'version': '3.0'}
 }
-
-
-class Vertex:
-    __slots__ = (
-        "co",
-        "normal",
-        "uvs",
-        "colors",
-        "loop_indices",
-        "index",
-        "weights",
-        "joint_indexes",
-        )
-
-    def __init__(self, mesh, loop):
-        vert_idx = loop.vertex_index
-        loop_idx = loop.index
-        self.co = mesh.vertices[vert_idx].co.freeze()
-        self.normal = loop.normal.freeze()
-        self.uvs = tuple(layer.data[loop_idx].uv.freeze() for layer in mesh.uv_layers)
-        self.colors = tuple(layer.data[loop_idx].color.freeze() for layer in mesh.vertex_colors)
-        self.loop_indices = [loop_idx]
-
-        # Take the four most influential groups
-        groups = sorted(
-            mesh.vertices[vert_idx].groups,
-            key=lambda group: group.weight,
-            reverse=True
-        )
-        if len(groups) > 4:
-            groups = groups[:4]
-
-        self.weights = [group.weight for group in groups]
-        self.joint_indexes = [group.group for group in groups]
-
-        if len(self.weights) < 4:
-            for _ in range(len(self.weights), 4):
-                self.weights.append(0.0)
-                self.joint_indexes.append(0)
-
-        self.index = 0
-
-    def __hash__(self):
-        return hash((self.co, self.normal, self.uvs, self.colors))
-
-    def __eq__(self, other):
-        equals = (
-            (self.co == other.co) and
-            (self.normal == other.normal) and
-            (self.uvs == other.uvs) and
-            (self.colors == other.colors)
-        )
-
-        if equals:
-            indices = self.loop_indices + other.loop_indices
-            self.loop_indices = indices
-            other.loop_indices = indices
-        return equals
-
-
-class Buffer:
-    ARRAY_BUFFER = 34962
-    ELEMENT_ARRAY_BUFFER = 34963
-
-    BYTE = 5120
-    UNSIGNED_BYTE = 5121
-    SHORT = 5122
-    UNSIGNED_SHORT = 5123
-    INT = 5124
-    UNSIGNED_INT = 5125
-
-    FLOAT = 5126
-
-    MAT4 = 'MAT4'
-    VEC4 = 'VEC4'
-    VEC3 = 'VEC3'
-    VEC2 = 'VEC2'
-    SCALAR = 'SCALAR'
-
-    class Accessor:
-        __slots__ = (
-            "name",
-            "buffer",
-            "buffer_view",
-            "byte_offset",
-            "byte_stride",
-            "component_type",
-            "count",
-            "min",
-            "max",
-            "data_type",
-            "type_size",
-            "_ctype",
-            "_ctype_size",
-            "_buffer_data",
-            )
-
-        def __init__(self,
-                     name,
-                     buffer,
-                     buffer_view,
-                     byte_offset,
-                     byte_stride,
-                     component_type,
-                     count,
-                     data_type):
-            self.name = name
-            self.buffer = buffer
-            self.buffer_view = buffer_view
-            self.byte_offset = byte_offset
-            self.byte_stride = byte_stride
-            self.component_type = component_type
-            self.count = count
-            self.min = [math.inf for i in range(16)]
-            self.max = [0 for i in range(16)]
-            self.data_type = data_type
-
-            if self.data_type == Buffer.MAT4:
-                self.type_size = 16
-            elif self.data_type == Buffer.VEC4:
-                self.type_size = 4
-            elif self.data_type == Buffer.VEC3:
-                self.type_size = 3
-            elif self.data_type == Buffer.VEC2:
-                self.type_size = 2
-            else:
-                self.type_size = 1
-
-            if component_type == Buffer.BYTE:
-                self._ctype = '<b'
-            elif component_type == Buffer.UNSIGNED_BYTE:
-                self._ctype = '<B'
-            elif component_type == Buffer.SHORT:
-                self._ctype = '<h'
-            elif component_type == Buffer.UNSIGNED_SHORT:
-                self._ctype = '<H'
-            elif component_type == Buffer.INT:
-                self._ctype = '<i'
-            elif component_type == Buffer.UNSIGNED_INT:
-                self._ctype = '<I'
-            elif component_type == Buffer.FLOAT:
-                self._ctype = '<f'
-            else:
-                raise ValueError("Bad component type")
-
-            self._ctype_size = struct.calcsize(self._ctype)
-            self._buffer_data = self.buffer.get_buffer_data(self.buffer_view)
-
-        def __len__(self):
-            return self.count
-
-        def __getitem__(self, idx):
-            if not isinstance(idx, int):
-                raise TypeError("Expected an integer index")
-
-            ptr = (
-                (
-                    (idx % self.type_size)
-                    * self._ctype_size + idx // self.type_size * self.byte_stride
-                ) + self.byte_offset
-            )
-
-            return struct.unpack_from(self._ctype, self._buffer_data, ptr)[0]
-
-        def __setitem__(self, idx, value):
-            if not isinstance(idx, int):
-                raise TypeError("Expected an integer index")
-
-            i = idx % self.type_size
-            self.min[i] = value if value < self.min[i] else self.min[i]
-            self.max[i] = value if value > self.max[i] else self.max[i]
-
-            ptr = (
-                (i * self._ctype_size + idx // self.type_size * self.byte_stride)
-                + self.byte_offset
-            )
-
-            struct.pack_into(self._ctype, self._buffer_data, ptr, value)
-
-    __slots__ = (
-        "name",
-        "buffer_type",
-        "bytelength",
-        "buffer_views",
-        "accessors",
-        )
-
-    def __init__(self, name):
-        self.name = 'buffer_{}'.format(name)
-        self.buffer_type = 'arraybuffer'
-        self.bytelength = 0
-        self.buffer_views = collections.OrderedDict()
-        self.accessors = {}
-
-    def export_buffer(self, state):
-        data = bytearray()
-        for view in self.buffer_views.values():
-            data.extend(view['data'])
-
-        if state['settings']['buffers_embed_data']:
-            uri = 'data:text/plain;base64,' + base64.b64encode(data).decode('ascii')
-        else:
-            uri = bpy.path.clean_name(self.name) + '.bin'
-            with open(os.path.join(state['settings']['gltf_output_dir'], uri), 'wb') as fout:
-                fout.write(data)
-
-        return {
-            'byteLength': self.bytelength,
-            'type': self.buffer_type,
-            'uri': uri,
-        }
-
-    def add_view(self, bytelength, target):
-        buffer_name = 'bufferView_{}_{}'.format(self.name, len(self.buffer_views))
-        self.buffer_views[buffer_name] = {
-            'data': bytearray(bytelength),
-            'target': target,
-            'bytelength': bytelength,
-            'byteoffset': self.bytelength,
-        }
-        self.bytelength += bytelength
-        return buffer_name
-
-    def export_views(self):
-        gltf = {}
-
-        for key, value in self.buffer_views.items():
-            gltf[key] = {
-                'buffer': self.name,
-                'byteLength': value['bytelength'],
-                'byteOffset': value['byteoffset'],
-            }
-
-            if value['target'] is not None:
-                gltf[key]['target'] = value['target']
-
-        return gltf
-
-    def get_buffer_data(self, buffer_view):
-        return self.buffer_views[buffer_view]['data']
-
-    def add_accessor(self,
-                     buffer_view,
-                     byte_offset,
-                     byte_stride,
-                     component_type,
-                     count,
-                     data_type):
-        accessor_name = 'accessor_{}_{}'.format(self.name, len(self.accessors))
-        self.accessors[accessor_name] = self.Accessor(
-            accessor_name,
-            self,
-            buffer_view,
-            byte_offset,
-            byte_stride,
-            component_type,
-            count,
-            data_type
-        )
-        return self.accessors[accessor_name]
-
-    def export_accessors(self):
-        gltf = {}
-
-        for key, value in self.accessors.items():
-            # Do not export an empty accessor
-            if value.count == 0:
-                continue
-
-            gltf[key] = {
-                'bufferView': value.buffer_view,
-                'byteOffset': value.byte_offset,
-                'byteStride': value.byte_stride,
-                'componentType': value.component_type,
-                'count': value.count,
-                'min': value.min[:value.type_size],
-                'max': value.max[:value.type_size],
-                'type': value.data_type,
-            }
-
-        return gltf
-
-    def __add__(self, other):
-        # Handle the simple stuff
-        combined = Buffer('combined')
-        combined.bytelength = self.bytelength + other.bytelength
-        combined.accessors = {**self.accessors, **other.accessors}
-
-        # Need to update byte offsets in buffer views
-        combined.buffer_views = self.buffer_views.copy()
-        other_views = other.buffer_views.copy()
-        for key in other_views.keys():
-            other_views[key]['byteoffset'] += self.bytelength
-        combined.buffer_views.update(other_views)
-
-        return combined
 
 
 def togl(matrix):
