+++ types/Vertex.py
@@ -1,0 +1,56 @@
+class Vertex:
+	__slots__ = (
+		"co",
+		"normal",
+		"uvs",
+		"colors",
+		"loop_indices",
+		"index",
+		"weights",
+		"joint_indexes",
+		)
+
+	def __init__(self, mesh, loop):
+		vert_idx = loop.vertex_index
+		loop_idx = loop.index
+		self.co = mesh.vertices[vert_idx].co.freeze()
+		self.normal = loop.normal.freeze()
+		self.uvs = tuple(layer.data[loop_idx].uv.freeze() for layer in mesh.uv_layers)
+		self.colors = tuple(layer.data[loop_idx].color.freeze() for layer in mesh.vertex_colors)
+		self.loop_indices = [loop_idx]
+
+		# Take the four most influential groups
+		groups = sorted(
+			mesh.vertices[vert_idx].groups,
+			key=lambda group: group.weight,
+			reverse=True
+		)
+		if len(groups) > 4:
+			groups = groups[:4]
+
+		self.weights = [group.weight for group in groups]
+		self.joint_indexes = [group.group for group in groups]
+
+		if len(self.weights) < 4:
+			for _ in range(len(self.weights), 4):
+				self.weights.append(0.0)
+				self.joint_indexes.append(0)
+
+		self.index = 0
+
+	def __hash__(self):
+		return hash((self.co, self.normal, self.uvs, self.colors))
+
+	def __eq__(self, other):
+		equals = (
+			(self.co == other.co) and
+			(self.normal == other.normal) and
+			(self.uvs == other.uvs) and
+			(self.colors == other.colors)
+		)
+
+		if equals:
+			indices = self.loop_indices + other.loop_indices
+			self.loop_indices = indices
+			other.loop_indices = indices
+		return equals
